Moving `Chromaticity` to submodule XYZ::chromaticity.rs
