+++ src/xyz.rs
@@ -32,56 +32,16 @@
 //!    (e.g., non-negative, within spectral locus) may be added in future refinements.  
 //!  
 
+mod chromaticity;
 mod dominant;
+pub use chromaticity::Chromaticity;
 
 use core::f64;
 
 use crate::{error::Error, observer::Observer, rgb::RgbSpace, rgb::WideRgb};
 use approx::AbsDiffEq;
-use nalgebra::{ArrayStorage, Vector2, Vector3};
+use nalgebra::{ArrayStorage, Vector3};
 use wasm_bindgen::prelude::wasm_bindgen;
-
-/// A chromaticity coordinate with x and y values.
-#[wasm_bindgen]
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct Chromaticity {
-    xy: Vector2<f64>,
-}
-
-impl Chromaticity {
-    /// Returns a new `Chromaticity` object with the given x and y coordinates.
-    pub const fn new(x: f64, y: f64) -> Self {
-        let xy = Vector2::new(x, y);
-        Self { xy }
-    }
-
-    /// Returns the x coordinate.
-    pub fn x(self) -> f64 {
-        self.xy.x
-    }
-
-    /// Returns the y coordinate.
-    pub fn y(self) -> f64 {
-        self.xy.y
-    }
-
-    /// Returns the chromaticity coordinates as an array in the format [x, y].
-    ///
-    /// ```
-    /// # use colorimetry::xyz::Chromaticity;
-    /// let [x, y] = [0.3127, 0.3290];
-    /// let chromaticity = Chromaticity::new(x, y);
-    /// assert_eq!(chromaticity.to_array(), [x, y]);
-    /// ```
-    pub fn to_array(self) -> [f64; 2] {
-        *self.xy.as_ref()
-    }
-
-    /// Returns the chromaticity coordinate as an `nalgebra` vector.
-    pub const fn to_vector(self) -> Vector2<f64> {
-        self.xy
-    }
-}
 
 #[wasm_bindgen]
 #[derive(Clone, Copy, Debug, PartialEq, Default)]
